+++ packages/stream_chat_flutter/lib/src/utils/utils.dart
@@ -2,4 +2,5 @@ export 'device_segmentation.dart';
 export 'extensions.dart';
 export 'helpers.dart';
 export 'jitter.dart';
+export 'stream_image_cdn.dart';
 export 'typedefs.dart';