+++ packages/stream_chat_flutter/test/src/attachment/image_attachment_test.dart
@@ -20,25 +20,28 @@ void main() {
 
       await tester.pumpWidget(
         MaterialApp(
-          home: StreamChatTheme(
-            data: streamTheme,
-            child: StreamChannel(
-              channel: channel,
-              child: SizedBox(
-                child: StreamImageAttachment(
-                  constraints: BoxConstraints.tight(const Size(
-                    300,
-                    300,
-                  )),
-                  message: Message(),
-                  image: Attachment(
-                    type: 'image',
-                    title: 'example.png',
-                    imageUrl:
-                        'https://logowik.com/content/uploads/images/flutter5786.jpg',
-                    extraData: const {
-                      'mime_type': 'png',
-                    },
+          home: StreamChatConfiguration(
+            data: StreamChatConfigurationData(),
+            child: StreamChatTheme(
+              data: streamTheme,
+              child: StreamChannel(
+                channel: channel,
+                child: SizedBox(
+                  child: StreamImageAttachment(
+                    constraints: BoxConstraints.tight(const Size(
+                      300,
+                      300,
+                    )),
+                    message: Message(),
+                    image: Attachment(
+                      type: 'image',
+                      title: 'example.png',
+                      imageUrl:
+                          'https://logowik.com/content/uploads/images/flutter5786.jpg',
+                      extraData: const {
+                        'mime_type': 'png',
+                      },
+                    ),
                   ),
                 ),
               ),