+++ packages/stream_chat_flutter/test/src/attachment/url_attachment_test.dart
@@ -19,18 +19,21 @@ void main() {
 
       await tester.pumpWidget(
         MaterialApp(
-          home: StreamChatTheme(
-            data: streamTheme,
-            child: StreamChannel(
-              channel: channel,
-              child: SizedBox(
-                child: StreamUrlAttachment(
-                  messageTheme: streamTheme.ownMessageTheme,
-                  message: Message(),
-                  hostDisplayName: 'Test',
-                  urlAttachment: Attachment(
-                    title: 'Flutter',
-                    titleLink: 'https://flutter.dev',
+          home: StreamChatConfiguration(
+            data: StreamChatConfigurationData(),
+            child: StreamChatTheme(
+              data: streamTheme,
+              child: StreamChannel(
+                channel: channel,
+                child: SizedBox(
+                  child: StreamUrlAttachment(
+                    messageTheme: streamTheme.ownMessageTheme,
+                    message: Message(),
+                    hostDisplayName: 'Test',
+                    urlAttachment: Attachment(
+                      title: 'Flutter',
+                      titleLink: 'https://flutter.dev',
+                    ),
                   ),
                 ),
               ),